refactoring

+++ Renderer/SSDX11CubeScene.cpp
@@ -9,7 +9,7 @@ void SSDX11CubeScene::InitializeScene()
 	mCube = std::make_shared<SSCube>();
 	mCube->SetPositionX(10);
 
-	mSphere = std::make_shared<SSSphere>(20,20,10);
+	mSphere = std::make_shared<SSSphere>(static_cast<UINT>(20),static_cast<UINT>(20), 10.0f);
 	mSphere->SetPositionX(10);
 	mSphere->SetPositionY( 10);
